A plethora of changes for Minecraft PE 0.11
Fix attackHook
Fix Custom armour
Disable screen-related methods
Fix forgetting to use marauder translation function in item ID patching
comment out most stuff that's causing dlsym errors
fix custom entity renderers
Enable texture packs when imported

+++ src/net/zhuoweizhang/mcpelauncher/ScriptManager.java
@@ -343,7 +343,7 @@ public class ScriptManager {
 		serverPort = 0;
 	}
 
-	public static void attackCallback(int attacker, int victim) {
+	public static void attackCallback(long attacker, long victim) {
 		callScriptMethod("attackHook", attacker, victim);
 	}
